Fix bug
Currently, when multiple film pairs are attempting to be retrieved
from the database, Array(numberOfPairs).fill() is actually executing
the function argument once and using this single return value in all
indexes of the array. Change things so the function argument in fill
is executed once for each index rather than once overall.